test LayerShuffle
